+++ helpers/mongo/schema/news-sources/whitelist-schema.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 
 module.exports = {
-  whiteListStructureSchema = new mongoose.Schema({
+  whitelistSourcesSchema = new mongoose.Schema({
     topic: { type: String, default: null },
     domains: [String]
   })
